Restructured test directory
