+++ api/src/test/java/com/codeforcommunity/rest/ApiRouterTest.java
@@ -1,37 +1,17 @@
-package com.codeforcommunity.test;
-
-import com.codeforcommunity.rest.ApiRouter;
-
-import java.lang.ProcessBuilder;
-import java.io.File;
-import java.util.Arrays;
-import java.net.HttpURLConnection;
-import java.io.IOException;
+package com.codeforcommunity.rest;
 
 import io.vertx.core.Vertx;
 import io.vertx.core.http.*;
-import io.vertx.core.DeploymentOptions;
-import io.vertx.core.VertxOptions;
-import io.vertx.core.json.*;
-
 import io.vertx.ext.web.*;
-import io.vertx.ext.unit.TestSuite;
-import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
-import io.vertx.ext.web.Route;
 import io.vertx.ext.web.RoutingContext;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
 @RunWith(VertxUnitRunner.class)
